backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
pycryptodome==3.19.0
qrcode==7.3
Pillow==10.0.0
